+++ android/versioned-abis/expoview-abi39_0_0/src/main/java/abi39_0_0/org/unimodules/interfaces/sensors/services/MagnetometerService.java
@@ -1,6 +1,0 @@
-package abi39_0_0.org.unimodules.interfaces.sensors.services;
-
-import abi39_0_0.org.unimodules.interfaces.sensors.SensorService;
-
-public interface MagnetometerService extends SensorService {
-}