+++ docs/scripts/test-links.js
@@ -1,6 +1,12 @@
 const puppeteer = require('puppeteer');
 
 const url = process.argv[2];
+
+if (!url) {
+  console.error(`You need to provide the base URL for links as a parameter.`);
+  process.exit(-1);
+}
+
 const externalLinks = [
   '/versions/latest/workflow/expo-cli/', // https://github.com/expo/expo-cli/blob/master/packages/expo-cli/README.md and https://github.com/expo/expo-cli/blob/master/README.md
   '/versions/latest/workflow/configuration/', // https://github.com/expo/expo-cli/blob/master/CONTRIBUTING.md and https://github.com/expo/expo-cli/blob/master/packages/expo-cli/src/commands/init.js and https://github.com/expo/expo-cli/blob/master/packages/xdl/src/project/Doctor.js
@@ -31,25 +37,18 @@ const externalLinks = [
     for (const link of externalLinks) {
       const response = await page.goto(`${url}${link}`);
       if (response.status() === 404) {
-        await page.waitFor(
-          () => {
-            return (
-              document.querySelector('#redirect-link') || document.querySelector('#__not_found')
-            );
-          },
+        await page.waitForFunction(
+          () => document.querySelector('#redirect-link') || document.querySelector('#__not_found'),
           { timeout: 500 }
         );
         if (await page.$('#redirect-link')) {
           await Promise.all([page.waitForNavigation(), page.click('#redirect-link')]);
           console.info(`Redirected from ${link} to ${await page.url()}`);
           try {
-            await page.waitFor(
-              () => {
-                return (
-                  document.querySelector('#__redirect_failed') ||
-                  document.querySelector('#__not_found')
-                );
-              },
+            await page.waitForFunction(
+              () =>
+                document.querySelector('#__redirect_failed') ||
+                document.querySelector('#__not_found'),
               { timeout: 500 }
             );
             console.debug(`Redirect failed`);