+++ packages/expo-image/src/Image.tsx
@@ -33,6 +33,7 @@ export interface ImageProps extends AccessibilityProps {
   // or not.
   source?: ImageSourcePropType | null;
   style?: StyleProp<ImageStyle>;
+  defaultSource?: ImageSourcePropType | null;
   resizeMode?: ImageResizeMode;
   /**
    * @Android only