+++ packages/expo-modules-autolinking/src/types.ts
@@ -15,7 +15,6 @@ export interface SearchOptions {
 
 export interface ResolveOptions extends SearchOptions {
   json?: boolean;
-  packageListTarget?: string | null;
 }
 
 export interface GenerateOptions extends ResolveOptions {